'auth-repas'

+++ database/migrations/2023_12_23_234250_create_steroids_table.php
@@ -15,7 +15,7 @@ class CreateSteroidsTable extends Migration
     {
         Schema::create('steroids', function (Blueprint $table) {
             $table->id();
-            $table->tinyInteger('avg')->default(0);
+            $table->double('avg',8,2)->default(0);
             $table->timestamps();
         });
     }
